feat(demeaner): default MAP backend to rust
