+++ docs/_sidebar.yml
@@ -44,7 +44,7 @@ website:
       - reference/estimation.formula.factor_interaction.factor_interaction.qmd
       section: Formula Parsing & Model Matrix
     - contents:
-      - reference/estimation.internals.demean_.demean.qmd
+      - reference/estimation.demean.qmd
       - reference/core.detect_singletons.detect_singletons.qmd
       - reference/utils.utils.ssc.qmd
       - reference/utils.utils.get_ssc.qmd
