+++ pyfixest/estimation/internals/vcov_utils.py
@@ -1,4 +1,3 @@
-import numba as nb
 import numpy as np
 import pandas as pd
 
@@ -186,121 +185,3 @@ def _prepare_twoway_clustering(clustervar: list, cluster_df: pd.DataFrame):
     )
 
     return cluster_df
-
-
-# CODE from Styfen Schaer (@styfenschaer)
-@nb.njit(parallel=False)
-def bucket_argsort(arr: np.ndarray) -> tuple[np.ndarray, np.ndarray]:
-    """
-    Sorts the input array using the bucket sort algorithm.
-
-    Parameters
-    ----------
-    arr : array_like
-        An array_like object that needs to be sorted.
-
-    Returns
-    -------
-    array_like
-        A sorted copy of the input array.
-
-    Raises
-    ------
-    ValueError
-        If the input is not an array_like object.
-
-    Notes
-    -----
-    The bucket sort algorithm works by distributing the elements of an array
-    into a number of buckets. Each bucket is then sorted individually, either
-    using a different sorting algorithm, or by recursively applying the bucket
-    sorting algorithm.
-    """
-    counts = np.zeros(arr.max() + 1, dtype=np.uint32)
-    for i in range(arr.size):
-        counts[arr[i]] += 1
-
-    locs = np.empty(counts.size + 1, dtype=np.uint32)
-    locs[0] = 0
-    pos = np.empty(counts.size, dtype=np.uint32)
-    for i in range(counts.size):
-        locs[i + 1] = locs[i] + counts[i]
-        pos[i] = locs[i]
-
-    args = np.empty(arr.size, dtype=np.uint32)
-    for i in range(arr.size):
-        e = arr[i]
-        args[pos[e]] = i
-        pos[e] += 1
-
-    return args, locs
-
-
-# CODE from Styfen Schaer (@styfenschaer)
-@nb.njit(parallel=False)
-def _crv1_meat_loop(
-    scores: np.ndarray,
-    clustid: np.ndarray,
-    cluster_col: np.ndarray,
-) -> np.ndarray:
-    k = scores.shape[1]
-    dtype = scores.dtype
-    meat = np.zeros((k, k), dtype=dtype)
-
-    g_indices, g_locs = bucket_argsort(cluster_col)
-
-    score_g = np.empty((k, 1), dtype=dtype)
-    meat_i = np.empty((k, k), dtype=dtype)
-
-    for i in range(clustid.size):
-        g = clustid[i]
-        start = g_locs[g]
-        end = g_locs[g + 1]
-        g_index = g_indices[start:end]
-        score_g = scores[g_index, :].sum(axis=0)
-        np.outer(score_g, score_g, out=meat_i)
-        meat += meat_i
-
-    return meat
-
-
-@nb.njit(parallel=False)
-def _crv1_vcov_loop(
-    X: np.ndarray,
-    clustid: np.ndarray,
-    cluster_col: np.ndarray,
-    q: float,
-    u_hat: np.ndarray,
-    delta: float,
-) -> tuple[np.ndarray, np.ndarray]:
-    _, k = X.shape
-
-    A = np.zeros((k, k))
-    B = np.zeros((k, k))
-    g_indices, g_locs = bucket_argsort(cluster_col)
-
-    eps = 1e-7
-
-    for g in clustid:
-        start = g_locs[g]
-        end = g_locs[g + 1]
-        g_index = g_indices[start:end]
-
-        Xg = X[g_index, :]
-        ug = u_hat[g_index]
-
-        ng = g_index.size
-        for i in range(ng):
-            Xgi = Xg[i, :]
-            psi_i = q - 1.0 * (ug[i] <= eps)
-            for j in range(ng):
-                Xgj = Xg[j, :]
-                psi_j = q - 1.0 * (ug[j] <= eps)
-                A += np.outer(Xgi, Xgj) * psi_i * psi_j
-
-            mask_i = (np.abs(ug[i]) < delta) * 1.0
-            B += np.outer(Xgi, Xgi) * mask_i
-
-    B /= 2 * delta
-
-    return A, B
