+++ pyproject.toml
@@ -40,7 +40,6 @@ dependencies = [
     "joblib>=1.4.2",
     "maketables>=0.1.0",
     "narwhals>=1.13.3",
-    "numba>=0.58.0",
     "numpy>=1.25.2",
     "pandas>=1.1.0",
     "scipy>=1.6",
@@ -50,6 +49,9 @@ dependencies = [
 ]
 
 [project.optional-dependencies]
+numba = [
+    "numba>=0.58.0",
+]
 jax = [
     "jax>=0.4.15",
 ]
@@ -203,7 +205,7 @@ doubleml = ">=0.10.1"
 pytest-sugar = ">=1.1.1"
 
 [tool.pixi.pypi-dependencies]
-pyfixest = { path = ".", editable = true, extras = ["jax", "plots"] }
+pyfixest = { path = ".", editable = true, extras = ["jax", "plots", "numba"] }
 wildboottest = ">=0.3.2"
 maturin-import-hook = ">=0.2.0"
 maturin = ">=1.8"
